Moved definitions around
